tools/agents/agent-listener/requirements.txt
# ==============================================================================
# CORPORACAO SENCIENTE - Agent Listener Dependencies
# ==============================================================================

# Socket.IO client
python-socketio[asyncio_client]>=5.10.0
aiohttp>=3.9.0

# System metrics
psutil>=5.9.0

# Settings
pydantic-settings>=2.1.0

# Logging
structlog>=24.1.0

# Screenshot (opcional)
pillow>=10.0.0
pyautogui>=0.9.54
